+++ api-builder-plugin-fn-redis/test/test-config/basic-config.js
@@ -1,8 +1,0 @@
-module.exports = {
-	pluginConfig: {
-		'@axway-api-builder-ext/api-builder-plugin-fn-redis': {
-			host: 'api-env', 
-			port: '16379'
-		}
-	}
-};